pyproject.toml
[project]
name = "examquestionverification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agentscope>=1.0.6",
    "agentscope-runtime>=0.1.6",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
]
